support decimal values in intense traffic generator #406

+++ src/udpflooding.h
@@ -19,7 +19,7 @@ class ThreadSender : public QThread
         qint64  getElapsedMS();
         QString ip;
         quint16 port;
-        int delay;
+        double delay;
         bool speedSendEnabled;
         QString ascii;
         QByteArray hex;
